techou/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-debug-toolbar==4.4.6
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.67.0
httplib2==0.22.0
idna==3.10
mysqlclient==2.2.7
numpy==1.24.4
oauthlib==3.2.2
pandas==2.0.3
pkg_resources==0.0.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
